requirements.txt
pypozyx
pyyaml
requests